backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-filter==24.3
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
